Old one is now VispyCompoundImageLayer
